Finish implementing Jameson's PR feedback:
He said that the paramater to `jl_gc_collect` is controlling what the
_sweep phase_ does, not what the mark phase does. So you have to do a
JL_GC_FULL sweep *first,* because once everything's been sweeped, we can
clear all the mark bits, so that the *next* GC mark phase will mark the
whole object graph.

So we need to call GC twice, and just hope that no one triggers GC in
between the first and the second call. 😅